requirements.txt
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.0.8
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
bs4==0.0.2
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.2.0
cookiecutter==2.1.1
coverage==6.5.0
crispy-bootstrap5==2024.2
cryptography==38.0.4
cycler==0.12.1
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
Django==4.2.13
django-allauth==0.51.0
django-bootstrap-v5==1.0.11
django-browser-reload==1.6.0
django-crispy-forms==2.1
django-debug-toolbar==4.2.0
django-extensions==3.2.1
django-sslserver==0.22
django-tailwind==3.4.0
django-widget-tweaks==1.4.12
djangorestframework==3.15.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==1.1.0
fastjsonschema==2.19.1
filelock==3.13.1
flake8==6.1.0
fonttools==4.48.1
fqdn==1.5.1
gunicorn==20.1.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
isoduration==20.11.0
jedi==0.18.1
Jinja2==3.1.2
jinja2-time==0.2.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
jupyterthemes==0.20.0
kiwisolver==1.4.5
lesscpy==0.15.1
MarkupSafe==2.1.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.5
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.23.3
oauthlib==3.2.2
openpyxl==3.0.10
overrides==7.7.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.1
parso==0.8.3
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.0.0
ply==3.11
prometheus-client==0.19.0
prompt-toolkit==3.0.31
psutil==5.9.2
pure-eval==0.2.2
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
Pygments==2.13.0
PyJWT==2.6.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-json-logger==2.0.7
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.2.1
pywin32==304
pywinpty==2.0.12
PyYAML==6.0
pyzmq==24.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.5.1
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tomli==2.0.1
tornado==6.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2022.6
uri-template==1.3.0
urllib3==1.26.13
virtualenv==20.24.7
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.2
